Story ac2. Changed NumericPrincipal to use an int instead of a long.

+++ projects/cadcAccessControl-Server/src/ca/nrc/cadc/ac/server/web/users/UserActionFactory.java
@@ -261,7 +261,7 @@ public abstract class UserActionFactory
         else if (idType.equalsIgnoreCase(IdentityType.CADC.getValue()))
         {
             return new User<NumericPrincipal>(new NumericPrincipal(
-                    Long.parseLong(userName)));
+                    Integer.parseInt(userName)));
         }
         else if (idType.equalsIgnoreCase(IdentityType.OPENID.getValue()))
         {
